another float fix

+++ functions.cpp
@@ -13,7 +13,7 @@ void initializeNeighbor(neighbor array[], int SIZE){
         float acres = ((rand()%(32))+19);
         int garages = (acres/10);
         int garagesBuilt = 0;
-        int stolenAcres = 0;
+        float stolenAcres = 0;
         bool good = (((rand() % 10) + 1) <= 5)? true : false;
         bool active = true;
         neighbor neighbors = {ID, acres, garages, garagesBuilt, stolenAcres, good, active};
